requirements.txt
pymupdf>=1.18.10
numpy>=1.21.0
tqdm>=4.61.2
Pillow>=8.3.1
PyInstaller>=4.5.1